+++ src/main/java/HourlyData.java
@@ -3,18 +3,18 @@ import java.util.Map;
 public class HourlyData {
 
     private String time;
-    private Map<String, Double> windDirection;
+    private Map<String, Integer> windDirection;
     private Map<String, Double> windSpeed;
 
     public String getTime() {
         return time;
     }
 
-    public Map<String, Double> getWindDirection() {
+    public Map<String, Integer> getWindDirection() {
         return windDirection;
     }
 
-    public Double getWindDirectionValue(String key) {
+    public Integer getWindDirectionValue(String key) {
         return windDirection.get(key);
     }
